🦄 refactor(dailynote): 将部分dn相关代码抽离
