+++ src/types/index.d.ts
@@ -86,7 +86,7 @@ type SettingKey = (
     'PluginVersion' | "EnableMove" | 'EnableReserve' | 
     "ExpandGutterMenu" | 'PopupReserveDialog' | 'ResvEmbedAt' |
     'RetvType' | 'EnableResvDock' | 'DisableAutoCreateOnMobile' |
-    'NotebookBlacklist'
+    'NotebookBlacklist' | 'AutoOpenAfterSync'
 );
 
 interface ISettingItem {
